+++ dubbo-common/src/main/java/org/apache/dubbo/common/utils/LogUtil.java
@@ -16,10 +16,12 @@
  */
 package org.apache.dubbo.common.utils;
 
-import org.apache.dubbo.common.logger.Logger;
-import org.apache.dubbo.common.logger.LoggerFactory;
 
-import org.apache.log4j.Level;
+
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.slf4j.event.Level;
 
 import java.util.Iterator;
 import java.util.List;
@@ -50,9 +52,7 @@ public class LogUtil {
         List<Log> logList = DubboAppender.logList;
         for (int i = 0; i < logList.size(); i++) {
             String logName = logList.get(i).getLogName();
-            if (logName.contains(expectedLogName)) {
-                count++;
-            }
+            if (logName.contains(expectedLogName)) count++;
         }
         return count;
     }
@@ -62,9 +62,7 @@ public class LogUtil {
         List<Log> logList = DubboAppender.logList;
         for (int i = 0; i < logList.size(); i++) {
             Level logLevel = logList.get(i).getLogLevel();
-            if (logLevel.equals(expectedLevel)) {
-                count++;
-            }
+            if (logLevel.equals(expectedLevel)) count++;
         }
         return count;
     }
@@ -74,9 +72,8 @@ public class LogUtil {
         List<Log> logList = DubboAppender.logList;
         for (int i = 0; i < logList.size(); i++) {
             Log log = logList.get(i);
-            if (log.getLogLevel().equals(expectedLevel) && log.getLogThread().equals(threadName)) {
+            if (log.getLogLevel().equals(expectedLevel) && log.getLogThread().equals(threadName))
                 count++;
-            }
         }
         return count;
     }
@@ -86,9 +83,7 @@ public class LogUtil {
         List<Log> logList = DubboAppender.logList;
         for (int i = 0; i < logList.size(); i++) {
             String logThread = logList.get(i).getLogThread();
-            if (logThread.contains(expectedThread)) {
-                count++;
-            }
+            if (logThread.contains(expectedThread)) count++;
         }
         return count;
     }
@@ -98,9 +93,7 @@ public class LogUtil {
         List<Log> logList = DubboAppender.logList;
         for (int i = 0; i < logList.size(); i++) {
             String logMessage = logList.get(i).getLogMessage();
-            if (logMessage.contains(expectedMessage)) {
-                count++;
-            }
+            if (logMessage.contains(expectedMessage)) count++;
         }
         return count;
     }
@@ -112,9 +105,7 @@ public class LogUtil {
             Level logLevel = logList.get(i).getLogLevel();
             if (logLevel.equals(expectedLevel)) {
                 String logMessage = logList.get(i).getLogMessage();
-                if (logMessage.contains(expectedMessage)) {
-                    count++;
-                }
+                if (logMessage.contains(expectedMessage)) count++;
             }
         }
         return count;