+++ dubbo-common/src/test/java/org/apache/dubbo/common/logger/LoggerAdapterTest.java
@@ -20,13 +20,12 @@ import org.apache.dubbo.common.logger.jcl.JclLogger;
 import org.apache.dubbo.common.logger.jcl.JclLoggerAdapter;
 import org.apache.dubbo.common.logger.jdk.JdkLogger;
 import org.apache.dubbo.common.logger.jdk.JdkLoggerAdapter;
-import org.apache.dubbo.common.logger.log4j.Log4jLogger;
-import org.apache.dubbo.common.logger.log4j.Log4jLoggerAdapter;
 import org.apache.dubbo.common.logger.log4j2.Log4j2Logger;
 import org.apache.dubbo.common.logger.log4j2.Log4j2LoggerAdapter;
 import org.apache.dubbo.common.logger.slf4j.Slf4jLogger;
 import org.apache.dubbo.common.logger.slf4j.Slf4jLoggerAdapter;
 
+import org.apache.logging.slf4j.Log4jLogger;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
@@ -41,7 +40,6 @@ public class LoggerAdapterTest {
         return Stream.of(
                 Arguments.of(JclLoggerAdapter.class, JclLogger.class),
                 Arguments.of(JdkLoggerAdapter.class, JdkLogger.class),
-                Arguments.of(Log4jLoggerAdapter.class, Log4jLogger.class),
                 Arguments.of(Slf4jLoggerAdapter.class, Slf4jLogger.class),
                 Arguments.of(Log4j2LoggerAdapter.class, Log4j2Logger.class)
         );