+++ dubbo-common/src/test/java/org/apache/dubbo/common/logger/LoggerTest.java
@@ -18,7 +18,6 @@ package org.apache.dubbo.common.logger;
 
 import org.apache.dubbo.common.logger.jcl.JclLoggerAdapter;
 import org.apache.dubbo.common.logger.jdk.JdkLoggerAdapter;
-import org.apache.dubbo.common.logger.log4j.Log4jLoggerAdapter;
 import org.apache.dubbo.common.logger.log4j2.Log4j2LoggerAdapter;
 import org.apache.dubbo.common.logger.slf4j.Slf4jLoggerAdapter;
 
@@ -39,7 +38,6 @@ public class LoggerTest {
         return Stream.of(
                 Arguments.of(JclLoggerAdapter.class),
                 Arguments.of(JdkLoggerAdapter.class),
-                Arguments.of(Log4jLoggerAdapter.class),
                 Arguments.of(Slf4jLoggerAdapter.class),
                 Arguments.of(Log4j2LoggerAdapter.class)
         );