+++ dubbo-common/src/test/java/org/apache/dubbo/common/utils/LogUtilTest.java
@@ -17,9 +17,9 @@
 
 package org.apache.dubbo.common.utils;
 
-import org.apache.log4j.Level;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
+import org.slf4j.event.Level;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.Matchers.is;